@@ -4,18 +4,16 @@ import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import it.polimi.ingsw.Model.LeaderCard.LeaderCard;
 
-import java.io.BufferedReader;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
+import java.io.*;
 import java.lang.reflect.Type;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
 
 
 public class CardParser {
 
-    private static final String path = "src/main/resources/cards/LeaderCard.json";
+    private static final String path = "/cards/LeaderCard.json";
     private static final List<LeaderCardBuilder> LEADER_CARD_BUILDERS = parseLeaderDeckBuilder();
 
     public static List<LeaderCard> getLeaderCard(){
@@ -31,22 +29,15 @@ public class CardParser {
     private static List<LeaderCardBuilder> parseLeaderDeckBuilder() {
         Gson parser = new Gson();
         BufferedReader reader = null;
-        List<LeaderCardBuilder> cards = new ArrayList<>();
+        List<LeaderCardBuilder> cards;
 
         try {
-            reader = new BufferedReader(new FileReader(path));
+            reader = new BufferedReader(new InputStreamReader(it.polimi.ingsw.Parser.LeaderCardParser.CardParser.class.getResourceAsStream(path), StandardCharsets.UTF_8));
 
             Type collectionType = new TypeToken<List<LeaderCardBuilder>>(){}.getType();
-            List<LeaderCardBuilder> parsedCard = parser.fromJson(reader, collectionType);
 
-            if (parsedCard == null) {
-                //error
-            }
+            cards = parser.fromJson(reader, collectionType);
 
-            cards = parsedCard;
-
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
         } finally {
             if (reader != null) {
                 try {
